+++ embassy-net-esp-hosted/src/iface.rs
@@ -1,9 +1,13 @@
 //! Transport interface.
 //!
 //! The esp-hosted firmware supports multiple interfaces, selected at compile time.
-//! This module provides the [`Interface`] trait and implementations for SPI and HD-SPI.
+//! This module provides the [`Interface`] trait and implementations for SPI, HD-SPI
+//! and SDIO.
+
+use aligned::{A4, Aligned};
 
 pub mod hd_spi;
+pub mod sdio;
 pub mod spi;
 
 /// Physical interface trait for communicating with the ESP chip.
@@ -22,5 +26,5 @@ pub trait Interface {
     /// `tx_len` is the number of bytes in the payload.
     ///
     /// The payload bytes contain the valid length of the received buffer.
-    async fn transfer(&mut self, buffer: &mut [u8], tx_len: usize);
+    async fn transfer(&mut self, buffer: &mut Aligned<A4, [u8]>, tx_len: usize);
 }